Crate netscape

De/serialization via the Netscape cookie format

References

https://curl.se/docs/http-cookies.html

https://everything.curl.dev/http/cookies/fileformat.html

Structs

NetscapeCookieError

Functions

load

Load Netscape-cookie-formatted cookies from reader, skipping any expired cookies.

load_all

Load Netscape-cookie-formatted cookies from reader, loading both expired and unexpired cookies.

save

Serialize any unexpired and persistent cookies in the store to Netscape cookie format and write them to writer.

save_incl_expired_and_nonpersistent

Serialize all (including expired and non-persistent) cookies in the store to Netscape cookie format and write them to writer.

Type Aliases

StoreResult